correctly added discard modal when user makes changes in newprogram and navigates to the correct pages
