+++ client/src/components/programs/EditProgram.jsx
@@ -21,7 +21,6 @@ import { RoomInformation } from "./programComponents/RoomInformation"
 import { ProgramInformation } from "./programComponents/ProgramInformation"
 import { TimeFrequency } from "./programComponents/TimeFrequency"
 import { EmailDropdown } from "./programComponents/EmailDropdown";
-import { DeleteConfirmationModal } from "./DeleteConfirmationModal";
 
 export const EditProgram = () => {
   const { backend } = useBackendContext();
@@ -50,9 +49,6 @@ export const EditProgram = () => {
   const [bookingIds, setBookingIds] = useState([]);
   const [instructorSearchTerm, setInstructorSearchTerm] = useState("");
   const [payeeSearchTerm, setPayeeSearchTerm] = useState("");
-  const [hasChanges, setHasChanges] = useState(false);
-  const initialState = useRef(null);
-  const [isConfirmModalOpen, setIsConfirmModalOpen] = useState(false);
 
 
   useEffect(() => {
@@ -61,57 +57,6 @@ export const EditProgram = () => {
     getInitialAssignmentsData();
     getInitialLocations();
   }, []);
-
-  useEffect(() => {
-    // Store the initial state when the component first loads
-    initialState.current = {
-      eventName,
-      generalInformation,
-      selectedLocationId,
-      selectedInstructors,
-      selectedPayees,
-      repeatType,
-      repeatInterval,
-      startTime,
-      endTime,
-      startDate,
-      endDate,
-      selectedDays
-    };
-  }, []);
-
-  useEffect(() => {
-    // Compare the current state to the initial state
-    const isDifferent = JSON.stringify(initialState.current) !== JSON.stringify({
-      eventName,
-      generalInformation,
-      selectedLocationId,
-      selectedInstructors,
-      selectedPayees,
-      repeatType,
-      repeatInterval,
-      startTime,
-      endTime,
-      startDate,
-      endDate,
-      selectedDays
-    });
-
-    setHasChanges(isDifferent);
-  }, [
-    eventName,
-    generalInformation,
-    selectedLocationId,
-    selectedInstructors,
-    selectedPayees,
-    repeatType,
-    repeatInterval,
-    startTime,
-    endTime,
-    startDate,
-    endDate,
-    selectedDays
-  ]);
 
   useEffect(() => {
     getInstructorResults(instructorSearchTerm);
@@ -126,9 +71,6 @@ export const EditProgram = () => {
 }, [selectedLocationId]);
 
   const exit = () => {
-    if (hasChanges) {
-      setIsConfirmModalOpen(true);
-    }
     navigate('/programs/' + id);
   };
 
@@ -440,14 +382,6 @@ const payees = eventClientResponse.data
   return (
 
     <Navbar>
-      <DeleteConfirmationModal
-        isOpen={isConfirmModalOpen}
-        onClose={() => setIsConfirmModalOpen(false)}
-        onConfirm={() => {
-          setIsConfirmModalOpen(false); // Close modal
-          navigate('/programs/' + id); // Navigate if confirmed
-        }}
-      />
       <div id="body">
         <div id="programsBody">
           <div><Icon fontSize="2xl" onClick={exit} id="leftCancel"><IoCloseOutline/></Icon></div>
